project/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
duckduckgo-search==6.3.0
pypdf==5.0.1
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
